requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
supabase==1.0.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0.post2
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.2
pytest-asyncio==0.21.1 